+++ problema2/src/FloydWarshall.cpp
@@ -19,8 +19,8 @@ FloydWarshall::~FloydWarshall() {
  * de camino minimo. Las clases que implementen esta interfaz
  * deben implementar este metodo.
  */
-std::vector<std::vector<int>> FloydWarshall::resolver(std::vector<Ruta>& rutas, std::vector<int> costos, int n) {
-    std::vector<std::vector<int>> L = this->armarGrafoEnNiveles(rutas, costos, n);
+std::vector<std::vector<double>> FloydWarshall::resolver(std::vector<Ruta>& rutas, std::vector<int> costos, int n) {
+    std::vector<std::vector<double>> L = this->armarGrafoEnNiveles(rutas, costos, n);
     int N = n*61; // el tamaño de la matriz L es n*61, por los niveles
 
     for (int k = 0; k < N; ++k) {
