+++ problema2/src/FloydWarshall.h
@@ -20,7 +20,7 @@ public:
      * de camino minimo. Las clases que implementen esta interfaz
      * deben implementar este metodo.
      */
-    virtual std::vector<std::vector<int>> resolver(std::vector<Ruta>& rutas, std::vector<int> costos, int n);
+    virtual std::vector<std::vector<double>> resolver(std::vector<Ruta>& rutas, std::vector<int> costos, int n);
 
 };
